requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
python-multipart==0.0.6
numpy==1.26.4
matplotlib==3.8.2
seaborn==0.13.2
scikit-learn==1.5.1
sympy==1.12
pandas==2.2.2
jupyter==1.0.0
ipython==8.17.2
PyYAML==6.0.2